+++ scripts/release.sh
@@ -54,7 +54,9 @@ fi
 # Build Docker server image
 if [[ -z "$SKIP_DOCKER" ]]; then
   docker build -t $docker_name:$version .
+  docker build -t $docker_name:$version-standalone --build-arg STANDALONE=1 .
   docker tag $docker_name:$version $docker_name:latest
+  docker tag $docker_name:$version-standalone $docker_name:standalone
   # we shouldn't push docker this early in the script, but the docker image hash is not available until we do
   # and is needed for the SHA256SUMS file. https://groups.google.com/forum/#!topic/docker-user/PvAcxDrvP30
   # https://github.com/moby/moby/issues/16482 https://github.com/docker/distribution/issues/1662