+++ CHANGELOG.md
@@ -7,6 +7,9 @@
 
 - Add `--public-url` CLI option
 
+- Add standalone docker image variant available as
+  shesek/spark-wallet:x.y.z-standalone
+
 ## 0.2.0 - 2018-12-11
 
 - Channel management: display how much more we need to receive in order to spend